Moved request api tests from test_views.py to test_request_api.py
